+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## [11.0]
+
+- **Removed** `api.requestAll` action creator is now a private function.
+  Consumers _must_ use the method-specific API request action creators, which
+  now accept arrays of queries in addition to single query
+- **Removed** The `meta.promise` property of API actions has been removed - use
+  the `meta.request` Promise instead, which will return an array of responses.
+  E.g. instead of `api.post(...).meta.promise.then(response => ...)`, use
+  `api.post(...).meta.request.then(([response]) => ...)`
+
 ## [10.2]
 
 - **Removed** Oauth authentication for no-cookie requests. This means we no